fetch test refactor
divide to modules
