+++ miamoremustachio/package.json
@@ -3,8 +3,8 @@
   "version": "1.0.0",
   "main": "main.js",
   "dependencies": {
+    "@types/node": "^22.7.4"
   },
-  "devDependencies": {},
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
